requirements.txt
Flask==3.1.0
flask-cors==5.0.0
google-generativeai==0.8.3